% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_d}
\alias{calc_d}
\title{#' calcualte LR contribution
#' 
calculateContribution = function(TPM, )}
\usage{
calc_d(ydata)
}
\description{
#' calcualte LR contribution
#' 
calculateContribution = function(TPM, )
}
