% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_d_rcpp}
\alias{calc_d_rcpp}
\title{Calculate d using Rcpp}
\usage{
calc_d_rcpp(ydata)
}
\arguments{
\item{ydata}{A matrix of coordinates}

\item{v2}{Second value}
}
\value{
d
}
\description{
Calculate d using Rcpp
}
