% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcNormalizedConnection}
\alias{calcNormalizedConnection}
\title{Calculate normalized connection based on connection matrix and cell counts}
\usage{
calcNormalizedConnection(connection_mt, cell_count_table)
}
\arguments{
\item{connection_mt}{Named matrix}

\item{cell_count_table}{Cell count table generated by `table()` or a named vector recording the cell counts}
}
\value{
Named matrix of normalized connection
}
\description{
Calculate normalized connection based on connection matrix and cell counts
}
