% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combinePvalues}
\alias{combinePvalues}
\title{Combine p values
A wrapper function to combined p values from multiple indepedent tests
Require pacakge 'metap'}
\usage{
combinePvalues(p_values, method = "Stouffer", ...)
}
\arguments{
\item{p_values}{a vectors.}

\item{method}{One of the c("Fisher", "Pearson", "Edgington", "Tippett", "Lancaster", "Stouffer", "Truncated", "Wilkinson")}

\item{...}{Other arguments passed to metap functions}
}
\value{
A metap object with the combined p value
}
\description{
Combine p values
A wrapper function to combined p values from multiple indepedent tests
Require pacakge 'metap'
}
