% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAffinityMat}
\alias{getAffinityMat}
\title{Calculate affinity matrix}
\usage{
getAffinityMat(TPM, LR, denoise = 50, eps = 2.2251e-308, verbose = F, ...)
}
\arguments{
\item{TPM}{a TPM matrix with gene names as rownames and cell names as colnames}

\item{LR}{a dataframe/tibble record the information of ligand receptor pairs, 
have to have colnames "ligand", "receptor" and an optional third column with weights}

\item{denoise}{numeric value,}

\item{eps}{Minimum distances between cells}

\item{verbose}{logical. If TRUE, print out the progress information}
}
\description{
Calculate affinity matrix
}
