% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getContribution}
\alias{getContribution}
\title{get LR contribution for all the listed cluster pairs}
\usage{
getContribution(TPM, LR, detailed_connections)
}
\arguments{
\item{TPM}{a TPM matrix with gene names as rownames and cell names as colnames}

\item{LR}{a dataframe/tibble record the information of ligand receptor pairs, 
have to have colnames "ligand", "receptor" and an optional third column with weights}

\item{detailed_connections}{a list generated by `getSignificance`, 
which stored the connected cell pairs for each clutser pair}
}
\value{
a named list with sorted LR contributions
}
\description{
get LR contribution for all the listed cluster pairs
}
