% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Calculate 3D coordinates from expression}
\usage{
getCoordinates(TPM, LR, optimization = "Rcpp", verbose = F, ...)
}
\arguments{
\item{TPM}{a TPM matrix with gene names as rownames and cell names as colnames}

\item{LR}{a dataframe/tibble record the information of ligand receptor pairs, have to have colnames "ligand" and "receptor"}

\item{verbose}{logical. If TRUE, print out the progress information}

\item{...}{arguments passsed to optimization}

\item{optimiztion}{a string sepcify the optimization method to use. Can be one of 'origin.R','Rcpp', 'tSNE' or 'umap'.}
}
\description{
Calculate 3D coordinates from expression
}
