% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDensity3D}
\alias{getDensity3D}
\title{get 3D density}
\usage{
getDensity3D(x, y, z, n = 100, ...)
}
\arguments{
\item{x, y, z}{coordinates}

\item{n}{numbers of grid points to use for each dimension; recycled if length is less than 3.}
}
\description{
Wrapper for misc3d::kde3d to get density estimation
}
