% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSignificance}
\alias{getSignificance}
\title{Test for significance level}
\usage{
getSignificance(coordinates, labels, k = 3, verbose = F)
}
\arguments{
\item{labels}{a vector of celltype labels, correspond to the coordinates matrix}

\item{k}{a integer for top k connections}

\item{coordinate}{a 3D matrix}
}
\value{
A list contains coordinates, counts, p values and q values
}
\description{
Test for significance level
}
