% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labelData}
\alias{labelData}
\title{label data frame}
\format{
A data frame with 4645 rows and 2 variables:
\describe{
  \item{cells}{Cell names}
  \item{labels}{Cell labels}
}
}
\usage{
labelData
}
\description{
A data frame of the labels of cells, corresponding to cells in TPM.
}
\keyword{datasets}
