% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LR}
\alias{LR}
\title{LR data frame}
\format{
A data frame with 1961 rows and 3 variables:
\describe{
  \item{V1}{Gene names of ligands}
  \item{V2}{Gene names of receptors}
  \item{V3}{Weigths of the ligand-receptor pairs}
}

A data frame with 4645 rows and 2 variables:
\describe{
  \item{cells}{Cell names}
  \item{labels}{Cell labels}
}
}
\usage{
LR

LR
}
\description{
A data frame of 1961 ligand-receptor pairs.

A data frame of the labels of cells, corresponding to cells in TPM.
}
\keyword{datasets}
