% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilitatetSNE.R
\name{run_tSNE}
\alias{run_tSNE}
\title{Wrapper function for FItSNE: fast_tsne.R}
\usage{
run_tSNE(path2fast_tsneR = NULL, fast_tsne_path = NULL, verbose = T, ...)
}
\arguments{
\item{fast_tsne_path}{a string specify the path of executable binary fast_tsne}

\item{verbose}{Print running infos for debugging.}

\item{...}{include all the following fields that will be passed to fast_tsne}

\item{path2fast_tsne}{a string specify the fast_tsne.R from FIt-SNE}

\item{data_path}{a string specify the data_path passed to FIt-SNE}

\item{load_affinities}{If 'precomputed', input data X will be regarded as precomputed similarities and passed to fast_tsne
If 'load', input similarities will be loaded from the file.
If 'save', input similarities are saved into a file.
If 0 or NULL(default), affinities are neither saved nor loaded}

\item{X}{data matrix}

\item{dims}{dimensionality of the embedding. Default 2.}

\item{perplexity}{perplexity is used to determine the bandwidth of the Gaussian kernel in the input space. Default 30.}

\item{theta}{Set to 0 for exact.  If non-zero, then will use either Barnes Hut or FIt-SNE based on nbody_algo. If Barnes Hut, then
this determins the accuracy of BH approximation. Default 0.5.}

\item{max_iter}{Number of iterations of t-SNE to run. Default 1000.}

\item{ann_not_vptree}{use vp-trees (as in bhtsne) or approximate nearest neighbors (default).
set to be True for approximate nearest neighbors}

\item{exaggeration_factor}{coefficient for early exaggeration (>1). Default = 1, not used.}

\item{no_momentum_during_exag}{Set to 0 to use momentum and other optimization tricks. 1 to do plain,vanilla
gradient descent (useful for testing large exaggeration coefficients)}

\item{stop_early_exag_iter}{When to switch off early exaggeration. Default 250.}

\item{start_late_exag_iter}{When to start late exaggeration. 'auto' means that late exaggeration is not used, unless late_exag_coeff>0. In that
case, start_late_exag_iter is set to stop_early_exag_iter. Otherwise, set to equal the iteration at which late exaggeration should begin. Default 'auto'.}

\item{late_exag_coeff}{Late exaggeration coefficient. Set to -1 to not use late exaggeration. Default -1}

\item{learning_rate}{Set to desired learning rate or 'auto', which sets learning rate to N/exaggeration_factor where N is the sample size, or to 200 if
N/exaggeration_factor < 200. Default 'auto'}

\item{max_step_norm}{Maximum distance that a point is allowed to move on
one iteration. Larger steps are clipped to this value. This prevents                                                                                               
possible instabilities during gradient descent.  Set to -1 to switch it                                                                                            
off. Default: 5}

\item{nterms}{If using FIt-SNE, this is the number of interpolation points per sub-interval}

\item{intervals_per_integer}{See min_num_intervals}

\item{min_num_intervals}{Let maxloc = ceil(max(max(X))) and minloc = floor(min(min(X))). i.e. the points are in a [minloc]^no_dims by [maxloc]^no_dims interval/square.
The number of intervals in each dimension is either min_num_intervals or ceil((maxloc - minloc)/intervals_per_integer), whichever is                           
larger. min_num_intervals must be an integer >0, and intervals_per_integer must be >0. Default: min_num_intervals=50, intervals_per_integer = 1}

\item{sigma}{Fixed sigma value to use when perplexity==-1 Default -1 (None)}

\item{K}{Number of nearest neighbours to get when using fixed sigma Default -30 (None)}

\item{initialization}{'random', or N x no_dims array to intialize the solution. Default: 'random'.}

\item{perplexity_list}{if perplexity==0 then perplexity combination will
be used with values taken from perplexity_list. Default: NULL}

\item{df}{Degree of freedom of t-distribution, must be greater than 0.
Values smaller than 1 correspond to heavier tails, which can often                                                                                                 
resolve substructure in the embedding. See Kobak et al. (2019) for                                                                                                 
details. Default is 1.0}

\item{fft_not_bh}{logical, default FALSE, we only use tSNE for interpreting 3D coordinates}
}
\description{
Wrapper function for FItSNE: fast_tsne.R
}
