% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilitatetSNE.R
\name{runExactTSNE_R}
\alias{runExactTSNE_R}
\title{Run exact tsne, wrapper for integrated Exact TSNE calculation cpp}
\usage{
runExactTSNE_R(X, no_dims = 3, ...)
}
\arguments{
\item{X}{affinity matrix to input}

\item{no_dims}{integer; Output dimensionality; Default = 3}

\item{verbose}{logical; Whether to print out debug information; Default = TRUE}

\item{max_iter}{integer; maximum iteration; Default = 1000}

\item{Y_in}{user-defined intiate coordinates; Default = NULL}

\item{init}{TRUE if Y_in were specified}

\item{rand_seed}{integer; random seed default = -1, set by time.}

\item{max_step_norm}{Maximum distance that a point is allowed to move on one iteration. Larger steps are clipped to this value. 
This prevents possible instabilities during gradient descent. Set to -1 to switch it off. (Default: 5) #'}

\item{mom_switch_iter}{Numeric; (Default: 250)}

\item{momentum}{numeric; (Default 0.5)}

\item{final_momentum}{numeric; (Default 0.8)}

\item{df}{Degree of freedom of t-distribution, must be greater than 0. Values smaller than 1 correspond to heavier tails, 
which can often resolve substructure in the embedding. See Kobak et al. (2019) for details. Default is 1.0}
}
\description{
Run exact tsne, wrapper for integrated Exact TSNE calculation cpp
}
