% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TPM}
\alias{TPM}
\title{TPM matrix of 1000 genes with the expression of 23,686 genes}
\format{
A matrix with 23686 rows and 1000 variables:
}
\usage{
TPM
}
\description{
A dataset containing the expression data of 1000 cells.
}
\keyword{datasets}
