% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_grads_rcpp}
\alias{update_grads_rcpp}
\title{Update gradients using Rcpp}
\usage{
update_grads_rcpp(grads, ydata, stiffnesses)
}
\arguments{
\item{grads}{A matrix of gradients}

\item{ydata}{A matrix of coordinates}

\item{stiffnesses}{A matrix of stiffnesses}
}
\value{
updated grads
}
\description{
Update gradients using Rcpp
}
