% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAffinityMat}
\alias{getAffinityMat}
\title{Get affinityMatrix}
\usage{
getAffinityMat(TPM, LR, verbose = F, ...)
}
\arguments{
\item{TPM}{a TPM matrix with gene names as rownames and cell names as colnames}

\item{LR}{a dataframe/tibble record the information of ligand receptor pairs, have to have colnames "ligand" and "receptor"}

\item{verbose}{logical. If TRUE, print out the progress information}
}
\description{
Get affinityMatrix
}
