% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCoordinates}
\alias{getCoordinates}
\title{calculate coordinates}
\usage{
getCoordinates(TPM, LR, version = "cpp", verbose = F, ...)
}
\arguments{
\item{TPM}{a TPM matrix with gene names as rownames and cell names as colnames}

\item{LR}{a dataframe/tibble record the information of ligand receptor pairs, have to have colnames "ligand" and "receptor"}

\item{version}{The version of functions to run. 'origin' or 'cpp'}

\item{verbose}{logical. If TRUE, print out the progress information}

\item{...}{arguments passsed to optimization}
}
\description{
calculate coordinates
}
