% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot3D}
\alias{plot3D}
\title{Plot 3D figure using plotly
A wrapper function to plot 3D plots using plotly. Not necessary for CSOmap's core functions.}
\usage{
plot3D(
  plt_tbl,
  color_by = "density",
  title = "3D density",
  alpha = 0.8,
  save_path = NULL,
  ...
)
}
\arguments{
\item{plt_tbl}{data.frame/tibble; Should provide coordinates x,y,z.}

\item{color_by}{string; Specify that by which columns should the data points will be colored}

\item{title}{string; Title}

\item{alpha}{numeirc; 0-1 specify the alpha of dots}

\item{save_path}{string; Speicfy the saving path of the output 3D plot. a `/lib/` will also be generated with the output html. Default = NULL.}

\item{...}{Other arguments that will be passed to htmlwidgets::saveWidget}
}
\value{
a plotly object
}
\description{
Plot 3D figure using plotly
A wrapper function to plot 3D plots using plotly. Not necessary for CSOmap's core functions.
}
